+++ hpeOneView/oneview_client.py
@@ -111,6 +111,7 @@ from hpeOneView.resources.hypervisors.hypervisor_managers import HypervisorManag
 from hpeOneView.resources.security.certificates_server import CertificatesServer
 from hpeOneView.resources.hypervisors.hypervisor_cluster_profiles import HypervisorClusterProfiles
 from hpeOneView.resources.settings.appliance_configuration_timeconfig import ApplianceConfigurationTimeconfig
+from hpeOneView.resources.settings.appliance_ssh_access import ApplianceSshAccess
 
 ONEVIEW_CLIENT_INVALID_PROXY = 'Invalid Proxy format'
 ONEVIEW_CLIENT_INVALID_I3S_IP = 'Image streamer ip address is missing'
@@ -207,6 +208,7 @@ class OneViewClient(object):
         self.__hypervisor_managers = None
         self.__certificates_server = None
         self.__appliance_configuration_timeconfig = None
+        self.__appliance_ssh_access = None
 
     @classmethod
     def from_json_file(cls, file_name):
@@ -1214,3 +1216,13 @@ class OneViewClient(object):
             ApplianceConfigurationTimeconfig:
         """
         return ApplianceConfigurationTimeconfig(self.__connection)
+
+    @property
+    def appliance_ssh_access(self):
+        """
+        Gets the ApplianceSshAccess API client.
+
+        Returns:
+            ApplianceSshAccess:
+        """
+        return ApplianceSshAccess(self.__connection)