+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Extends support to below features for REST API version 2400.
 - Appliance Configuration Timeconfig
 - Appliance SNMPv3 Trap Destinations
 - Appliance Time and Locale Configuration
+- Appliance SSH Access
 
 # 5.6.0
 #### Notes